ADFCNN/requirements.txt
#Package                 Version
#----------------------- ------------
absl-py                 ==2.0.0
aiohttp                 ==3.8.5
aiosignal               ==1.3.1
async-timeout           ==4.0.3
attrs                   ==23.1.0
Braindecode             ==0.7
cachetools              ==5.3.1
certifi                 ==2023.7.22
charset-normalizer      ==3.2.0
contourpy               ==1.1.1
cycler                  ==0.11.0
decorator               ==5.1.1
easydict                ==1.10
fonttools               ==4.42.1
frozenlist              ==1.4.0
fsspec                  ==2023.9.2
google-auth             ==2.23.0
google-auth-oauthlib    ==1.0.0
grpcio                  ==1.58.0
h5py                    ==3.9.0
idna                    ==3.4
Jinja2                  ==3.1.2
joblib                  ==1.3.2
kiwisolver              ==1.4.5
lightning-utilities     ==0.9.0
Markdown                ==3.4.4
MarkupSafe              ==2.1.3
matplotlib              ==3.8.0
mne                     ==1.1.1
multidict               ==6.0.4
numpy                   ==1.25.0
oauthlib                ==3.2.2
packaging               ==23.1
pandas                  ==2.1.1
Pillow                  ==10.0.1
pip                     ==23.2.1
platformdirs            ==3.10.0
pooch                   ==1.7.0
protobuf                ==4.24.3
pyasn1                  ==0.5.0
pyasn1-modules          ==0.3.0
pyDeprecate             ==0.3.2
pyparsing               ==3.1.1
python-dateutil         ==2.8.2
pytorch-lightning       ==1.7.4
pytz                    ==2023.3.post1
PyYAML                  ==6.0.1
requests                ==2.31.0
requests-oauthlib       ==1.3.1
rsa                     ==4.9
scikit-learn            ==1.1.1
scipy                   ==1.10.0
setuptools              ==68.0.0
six                     ==1.16.0
skorch                  ==0.15.0
tabulate                ==0.9.0
tensorboard             ==2.14.0
tensorboard-data-server ==0.7.1
threadpoolctl           ==3.2.0
torch                   ==1.12.0+cu116
torchmetrics            ==0.9.3
torchsummary            ==1.5.1
tqdm                    ==4.66.1
typing_extensions       ==4.8.0
tzdata                  ==2023.3
urllib3                 ==1.26.16
Werkzeug                ==2.3.7
wheel                   ==0.38.4
yarl                    ==1.9.2
